api/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.9.4
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
